funcoes e recursos para projetos futuro

+++ --★theme★--/Update Spring 📦  2.6 To 3.0 ⭐/backend Spring v2.6.14/src/main/java/com/library/naldo/domain/Book.java
@@ -19,7 +19,7 @@ public class Book {
 	private String title;
 	private String author;
 	private String photo;
-	private Long isbn;
+	private Integer isbn;
 	private Double price;
 	private String language;
 	private String genre;
